refactor: pullToRefresh 를 페이지에 적용
